+++ src/web-request/land-req.js
@@ -1,10 +1,10 @@
 /*
 	This file is responsible for downloading the advertisement cases landing page.
-	Used for the 'read-lists' and 'download-pages' commands.
+	Used for the 'test-http', 'read-lists', and 'download-pages' commands.
 	When running 'download-pages', this is used to request the first page.
 */
 
-const asyncModule = require("async");
+
 const ora = require("ora");
 const linkPrep = require("../common/web/link-prep");
 const retryLimits = require("../common/web/retry-limits");