+++ stored-paths.js
@@ -6,6 +6,7 @@ const databaseFolderPath = path.join(".", "src", "common", "database", "scripts"
 const pageFolderPath = path.join(".", "pages");														// HTML pages folder.
 const reportsFolderPath = path.join(".", "reports");												// Advertisement case reports root folder.
 const exportFolderPath = path.join(".", "export");													// Default export folder.
+const testDownloadFolderPath = path.join(".", "test-download");										// Test download folder.
 const unknownOutputFilePath = path.join(".", "unknown-entries.csv");								// Default unknown entries file.
 
 const logFilePath = path.join(".", "log.txt");														// Log file path.
@@ -13,16 +14,21 @@ const dbCredsPath = path.join(".", "db-creds.json");												// Database cred
 const dbDefinitionPath = path.join(databaseFolderPath, "definition.sql");							// Definition script file.
 const dbClearAllPath = path.join(databaseFolderPath, "clear-all.sql");								// Clear all data script file.
 const dbClearCasesPath = path.join(databaseFolderPath, "clear-cases.sql");							// Clear existing cases script file.
+const testPagePath = path.join(testDownloadFolderPath, "page.html");								// Test HTML page download path.
+const testReportPath = path.join(testDownloadFolderPath, "report.pdf");								// Test report PDF download path.
 
 module.exports =
 {
 	pageFolder: pageFolderPath,
 	reportsFolder: reportsFolderPath,
 	exportFolder: exportFolderPath,
+	testDownloadFolder: testDownloadFolderPath,
 	unknownOutputFile: unknownOutputFilePath,
 	logFile: logFilePath,
 	dbCreds: dbCredsPath,
 	dbDefinition: dbDefinitionPath,
 	dbClearAll: dbClearAllPath,
-	dbClearCases: dbClearCasesPath
+	dbClearCases: dbClearCasesPath,
+	testPage: testPagePath,
+	testReport: testReportPath
 };